+++ test/sync/src/macosMain/kotlin/io/realm/mongodb/sync/SyncDispatcher.kt
@@ -1,0 +1,6 @@
+package io.realm.mongodb.sync
+
+import io.realm.internal.platform.singleThreadDispatcher
+import kotlinx.coroutines.CoroutineDispatcher
+
+actual fun getDispatcher(): CoroutineDispatcher = singleThreadDispatcher("Sync Dispatcher")